ref: simplify growingdagnetwork
